app_dir {binman}	R Documentation

Get application directory

Description

Get application directory

Usage

app_dir(appname, check = TRUE)

Arguments

appname
A character string giving the name of the application

check
check whether the app given by appname exists or not.

Value

A character string giving the path of the directory

Examples

## Not run: 
appdir <- app_dir("superduperapp", FALSE)

## End(Not run)

[Package binman version 0.1.1 Index]